TG: main.m - actually use full double precision for frametime; ODScene - likewise; ODPEntity - likewise; fix fallout following this changes

+++ AppsV2/OceanDemo/Entities/ODOceanEntity.m
@@ -340,7 +340,7 @@ static const NSUInteger defaultResolutionIndex = 0;
     return YES;
 }
 
-- (void) update:(const float)frameTime
+- (void) update:(const double)frameTime
 {
     [ projector     update:frameTime ];
     [ projectedGrid update:frameTime ];
