+++ AppsV2/OceanDemo/ODScene.h
@@ -73,7 +73,7 @@
 - (ODPreethamSkylight *) skylight;
 - (ODOceanEntity *) ocean;
 
-- (void) update:(const float)frameTime;
+- (void) update:(const double)frameTime;
 - (void) render;
 
 @end
